+++ sdks/browser4-python/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-All notable changes to browser4-sdk-python will be documented in this file.
+All notable changes to browser4-python will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.1.0] - TBD
 
 ### Added
-- Initial release of browser4-sdk-python
+- Initial release of browser4-python
 - Browser4Driver for automatic server management and lifecycle control
   - Automatic download of Browser4.jar from GitHub releases
   - Server startup, health checking, and graceful shutdown